% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{boolmat}
\alias{boolmat}
\title{boolmat}
\usage{
boolmat(gate, xp)
}
\arguments{
\item{gate}{A return from hypergate}

\item{xp}{Expression matrix as in the hypergate call}
}
\description{
Convert an expression matrix and a gating strategy to a boolean matrix (whether each event is gated out by each channel)
}
