% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{channels_contributions}
\alias{channels_contributions}
\title{channels_contributions}
\usage{
channels_contributions(gate, xp, gate_vector, level, beta = 1)
}
\arguments{
\item{gate}{A return from hypergate}

\item{xp}{Expression matrix as in the hypergate call}

\item{gate_vector}{Categorical vector of length nrow(xp)}

\item{level}{A level of gate_vector that identifies the population of interest}

\item{beta, }{should be the same as for the hypergate object}
}
\description{
Gives scores for the contribution of individual channels to a gating strategy
}
