% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{color_biplot_by_discrete}
\alias{color_biplot_by_discrete}
\title{Colors a biplot according to a vector with discrete values}
\usage{
color_biplot_by_discrete(matrix, discrete_vector, ..., bty = "l", pch = 16,
  cex = 0.5, colors = NULL, alpha = 100)
}
\arguments{
\item{matrix}{a two columns matrix}

\item{discrete_vector}{a vector of size nrow(matrix)}

\item{...}{passed to plot}
}
\description{
Colors a biplot according to a vector with discrete values
}
