% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{contract}
\alias{contract}
\title{contract}
\usage{
contract(par = par, xp_pos = envir$xp_pos, state_pos = envir$state_pos,
  xp_neg = envir$xp_neg, state_neg = envir$state_neg, n = envir$n,
  TP = envir$TP, TN = envir$TN, beta = envir$beta2,
  envir = parent.frame())
}
\description{
Test (some) possible contractions of the hyperrectangle
}
