% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{contract.update}
\alias{contract.update}
\title{contract.update}
\usage{
contract.update(contract_object, pars = envir$pars,
  active_channels = envir$active_channels, b_pos = envir$b_pos,
  b_neg = envir$b_neg, state_pos = envir$state_pos,
  state_neg = envir$state_neg, TN = envir$TN, TP = envir$TP,
  xp_pos = envir$xp_pos, xp_neg = envir$xp_neg, envir = parent.frame())
}
\description{
Update the hyperrectangle to the best contraction move found
}
