% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{coreloop}
\alias{coreloop}
\title{coreloop}
\usage{
coreloop(par, hg.env = hg.env$hg.env)
}
\arguments{
\item{par}{Current parametrization of the hyperrectangle}

\item{hg.env}{Environment where the main execution of hypergate takes place}
}
\description{
Core optimization loop of hypergate
}
