% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{expand}
\alias{expand}
\title{expand}
\usage{
expand(FN = envir$FN, FNTN_matrix = envir$FNTN_matrix, TP = envir$TP,
  TN = envir$TN, n = envir$n, beta = envir$beta2,
  envir = parent.frame())
}
\description{
Test (some) possible expansions of the hyperrectangle
}
