% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{expand.update}
\alias{expand.update}
\title{expand.update}
\usage{
expand.update(expand.object, pars = envir$pars, xp_pos = envir$xp_pos,
  xp_neg = envir$xp_neg, state_pos = envir$state_pos,
  state_neg = envir$state_neg, b_pos = envir$b_pos, b_neg = envir$b_neg,
  n = envir$n, TP = envir$TP, TN = envir$TN, envir = parent.frame())
}
\description{
Update the hyperrectangle to the best expansion move found
}
