% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{fill_FNTN_matrix}
\alias{fill_FNTN_matrix}
\title{fill_FNTN_matrix}
\usage{
fill_FNTN_matrix(xp_FN, xp_TN, B_FN, B_TN, par)
}
\arguments{
\item{xp_FN}{Expression matrix of False Negative events}

\item{xp_TN}{Expression matrix of True Negative events}

\item{B_FN}{Boolean matrix of FN events}

\item{B_TN}{Boolean matrix of TN events}

\item{par}{Current hyper-rectangle parametrization}
}
\description{
fill_FNTN_matrix Used for assessing whether an expansion move is possible
}
