% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{FNTN_matrix.recycle}
\alias{FNTN_matrix.recycle}
\title{FNTN_matrix.recycle}
\usage{
FNTN_matrix.recycle(FNTN_matrix, B_FN_old, B_TN_old, B_FN_new, B_TN_new, xp_FN,
  xp_TN, par)
}
\description{
Recycle an expansion matrix
}
