% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{hgate_info}
\alias{hgate_info}
\title{hgate_info}
\usage{
hgate_info(gate)
}
\arguments{
\item{gate}{A hypergate object (produced by hypergate())}
}
\value{
A data.frame with channel, sign, comp and threshold columns
}
\description{
Extract information about a hypergate return: the channels of the phenotype, the sign of the channels, the sign of the comparison, the thresholds.
}
\examples{
data(Samusik_01_subset)
xp=Samusik_01_subset$xp_src[,Samusik_01_subset$regular_channels]
gate_vector=Samusik_01_subset$labels
hg=hypergate(xp=xp,gate_vector=gate_vector,level=23,delta_add=0.01)
hgate_info(gate=hg)
hgate_pheno(gate=hg)
hgate_rule(gate=hg)
}
\seealso{
\code{hg_pheno}, \code{hg_rule}
}
