% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{hgate_pheno}
\alias{hgate_pheno}
\title{hgate_pheno}
\usage{
hgate_pheno(hgate, collapse = ", ")
}
\arguments{
\item{hgate}{A hypergate object (produced by hypergate())}

\item{collapse}{A character string to separate the markers.}
}
\value{
A string representing the phenotype.
}
\description{
Build a human readable phenotype, i.e. a combination of channels
  and sign (+ or -) from a hypergate return.
}
\examples{
## See hgate_info
}
\seealso{
\code{hg_rule}, \code{hg_info}
}
