% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{hgate_rule}
\alias{hgate_rule}
\title{hgate_info}
\usage{
hgate_rule(gate, collapse = ", ", digits = 2)
}
\arguments{
\item{gate}{A hypergate object (produced by hypergate())}
}
\value{
A data.frame with channel, sign, comp and threshold columns
}
\description{
Extract information about a hypergate return: the channels of the phenotype, the sign of the channels, the sign of the comparison, the thresholds.
}
\examples{
## See hgate_info
}
\seealso{
\code{\link{hg_pheno}}, \code{\link{hg_rule}}
}
