% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{hgate_sample}
\alias{hgate_sample}
\title{hgate_sample}
\usage{
hgate_sample(gate_vector, level, size = 1000, method = "prop")
}
\arguments{
\item{gate_vector}{A Categorical vector of length nrow(xp)}

\item{level}{A level of gate_vector so that gate_vector == level will produce
a boolean vector identifying events of interest}

\item{size}{An integer specifying the maximum number of events of interest to retain. If the count of events of interest is lower than \code{size}, than \code{size} will be set to that count.}

\item{method}{A string specifying the method to balance the count of events. \code{"prop"} means proportionnality: if events of interest are sampled in a 1/10 ratio, then all others events are sampled by the same ratio. \code{"10x"} means a balance of 10 between the count events of interest and the count all others events.}
}
\value{
A logical vector with TRUE correspond to the events being sampled
}
\description{
Downsample the data in order to fasten the computation and
  reduce the memory usage.
}
\note{
No replacement is applied. If there are less events in one group or the alternate than the algorithm requires, then all available events are returned.
}
\examples{
# Standard procedure with downsampling
data(Samusik_01_subset)
xp <- Samusik_01_subset$xp_src[,Samusik_01_subset$regular_channels]
gate_vector <- Samusik_01_subset$labels
sampled <- hgate_sample(gate_vector, level=8, 100)
table(sampled)
table(gate_vector[sampled])
xp_sampled <- xp[sampled, ]
gate_vector_sampled <- gate_vector[sampled]
hg <- hypergate(xp_sampled, gate_vector_sampled, level=8, delta_add=0.01)
# cluster 8 consists in 122 events
table(gate_vector)
# Downsampling
table(gate_vector[hgate_sample(gate_vector, level=8, 100)])
# Downsampling reduces the alternate events
table(gate_vector[hgate_sample(gate_vector, level=8, 100, "10x")])
# Downsampling is limited to the maximum number of events of interest
table(gate_vector[hgate_sample(gate_vector, level=8, 150)])
# Downsampling is limited to the maximum number of events of interest, and
# the alternate events are downsampled to a total of 10 times
table(gate_vector[hgate_sample(gate_vector, level=8, 150, "10x")])
}
