% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{makeTransparent}
\alias{makeTransparent}
\title{makeTransparent
Makes a color transparent}
\usage{
makeTransparent(colors, alpha = 255)
}
\arguments{
\item{colors}{A vector of colors as in `?col2rgb`}

\item{alpha}{transparency value (0=fully transparent, 255=fully opaque)}
}
\value{
A color in 9 characters format
}
\seealso{
http://stackoverflow.com/questions/8047668/transparent-equivalent-of-given-color
}
\author{
Ricardo Oliveros-Ramos
}
