% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{plot_gating_strategy}
\alias{plot_gating_strategy}
\title{plot_gating_strategy}
\usage{
plot_gating_strategy(gate, xp, gate_vector, level, highlight = "black",
  path = "./", ...)
}
\arguments{
\item{gate}{A hypergate object (produced by hypergate.stepwise.fullcycles())}

\item{xp}{The expression matrix from which the 'gate' parameter originates}

\item{gate_vector}{Categorical data from which the 'gate' parameter originates}

\item{level}{Level of gate_vector identifying the population of interest}

\item{highlight}{color of the positive population when plotting}

\item{path}{Where png files will be produced}

\item{...}{passed to png}
}
\description{
Plot a hypergate return
}
