% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{polygon.clean}
\alias{polygon.clean}
\title{Remove self intersection in polygons}
\usage{
polygon.clean(poly)
}
\arguments{
\item{poly}{a polygon (list with two components x and y which are equal-length numerical vectors)}
}
\value{
A polygon without overlapping edges and new vertices corresponding to non-inner points of intersection
}
\description{
Remove self intersection in polygons
}
