% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\docType{data}
\name{Samusik_01_subset}
\alias{Samusik_01_subset}
\title{2000 events randomly sampled from the 'Samusik_01' dataset}
\format{list with four elements: fs_src (a flowSet), xp_src (its expression matrix), labels (manual gates of the events) and tsne (a tSNE projection of the dataset)}
\usage{
Samusik_01_subset
}
\description{
2000 events randomly sampled from the 'Samusik_01' dataset
}
\references{
https://flowrepository.org/id/FR-FCM-ZZPH
}
\keyword{datasets}
