% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{subset_matrix_hg}
\alias{subset_matrix_hg}
\title{subset_matrix_hg}
\usage{
subset_matrix_hg(gate, xp)
}
\arguments{
\item{gate}{a return from hypergate}

\item{xp}{Expression matrix used for gate}
}
\description{
Returns a boolean vector whose TRUE elements correspond to events inside the hyperrectangle
}
