% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{subset_matrix_hyperrectangle.fast}
\alias{subset_matrix_hyperrectangle.fast}
\title{subset_matrix_hyperrectangle.fast}
\usage{
subset_matrix_hyperrectangle.fast(ui_x_xp, ci)
}
\description{
From a vector of boundary conditions and an expression matrix returns a boolean vector testing whether all conditions are satisfied
}
