% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{boolmat}
\alias{boolmat}
\title{boolmat}
\usage{
boolmat(gate, xp)
}
\arguments{
\item{gate}{A return from hypergate}

\item{xp}{Expression matrix as in the hypergate callxp=Samusik_01_subset$xp_src[,Samusik_01_subset$regular_channels]}
}
\description{
Convert an expression matrix and a gating strategy to a boolean matrix (whether each event is gated out by each channel)
}
\examples{
data(Samusik_01_subset)
xp=Samusik_01_subset$xp_src
gate_vector=Samusik_01_subset$labels
hg=hypergate(xp=xp,gate_vector=gate_vector,level=23,delta_add=0.01)
head(boolmat(hg,xp))
}
