% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{F_beta}
\alias{F_beta}
\title{F_beta}
\usage{
F_beta(pred, truth, beta = 1)
}
\arguments{
\item{pred}{boolean vector of predicted values}

\item{truth}{boolean vector of true values}
}
\description{
Compute a F_beta score comparing two boolean vectors
}
