% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergate.R
\name{hgate_rule}
\alias{hgate_rule}
\title{hgate_rule}
\usage{
hgate_rule(hgate, collapse = ", ", digits = 2)
}
\arguments{
\item{hgate}{A hypergate object (produced by hypergate())}

\item{collapse}{A character string to separate the markers.}

\item{digits}{An integer that specifies the decimal part when rounding.}
}
\value{
A data.frame with channel, sign, comp and threshold columns
}
\description{
Build a human readable rule i.e. a combination of channels, sign
  of comparison and threshold.
}
\examples{
## See hgate_info
}
\seealso{
\code{hg_pheno}, \code{hg_rule}
}
